moved raster related things out of geomesa-core.
