+++ geomesa-core/src/main/scala/org/locationtech/geomesa/core/index/Formatters.scala
@@ -113,24 +113,3 @@ case class CompositeTextFormatter(lf: Seq[TextFormatter], sep: String) extends T
   val numBits = lf.map(_.numBits).sum
   def formatString(gh: GeoHash, dt: DateTime, sf: SimpleFeature) = lf.map { _.formatString(gh, dt, sf) }.mkString(sep)
 }
-
-/**
- * Responsible for representing the resolution by encoding a double lexicographically.
- *
- * @param number
- */
-case class ScientificNotationTextFormatter(number: Double) extends TextFormatter {
-  val fmtdStr: String = lexiEncodeDoubleToString(number)
-  val numBits: Int = fmtdStr.length
-  def formatString(gh: GeoHash, dt: DateTime, sf: SimpleFeature) = fmtdStr
-}
-
-/**
- * Responsible for representing the Band Name of a given raster.
- *
- * @param bandName
- */
-case class RasterBandTextFormatter(bandName: String) extends TextFormatter {
-  val numBits: Int = bandName.length
-  def formatString(gh: GeoHash, dt: DateTime, sf: SimpleFeature) = bandName
-}
