+++ geomesa-core/src/main/scala/org/locationtech/geomesa/core/index/SchemaHelpers.scala
@@ -72,7 +72,6 @@ trait SchemaHelpers extends RegexParsers {
       case g: Geometry => innerSomewhere(g)
     }
 
-
   def pattern[T](p: => Parser[T], code: String): Parser[T] = CODE_START ~> p <~ (CODE_END + code)
 
   // A separator character, typically '%~#s' would indicate that elements are to be separated
@@ -125,19 +124,6 @@ trait SchemaHelpers extends RegexParsers {
 
   def datePlanner: Parser[DatePlanner] = datePattern ^^ {
     case fmt => DatePlanner(DateTimeFormat.forPattern(fmt))
-  }
-
-  // An Image Resolution encoder.
-  def resolutionPattern = pattern("[^%#]+".r, RESOLUTION_CODE)
-  def resolutionEncoder: Parser[ScientificNotationTextFormatter] = resolutionPattern ^^ {
-    case d => ScientificNotationTextFormatter(lexiDecodeStringToDouble(d))
-  }
-
-  // A Band encoder. 'RGB#b' would yield RGB
-  //  We match any string other that does *not* contain % or # since we use those for delimiters
-  def bandPattern = pattern("[^%#]+".r, BAND_CODE)
-  def bandEncoder: Parser[RasterBandTextFormatter] = bandPattern ^^ {
-    case b => RasterBandTextFormatter(b)
   }
 
   def geohashKeyPlanner: Parser[GeoHashKeyPlanner] = geohashPattern ^^ {
