+++ geomesa-core/src/test/scala/org/locationtech/geomesa/core/index/FormattersTest.scala
@@ -44,24 +44,4 @@ class FormattersTest extends Specification {
     }
   }
 
-  "ScientificNotationTextFormatter" should {
-    val testNumber = 123456789
-    val scientificNumberFormatter = ScientificNotationTextFormatter(testNumber)
-
-    "map a double to a string representation" in {
-      val scientificText = scientificNumberFormatter.formatString(null, null, null)
-      scientificText must be equalTo "c19d6f3454000000"
-    }
-  }
-
-  "RasterBandTextFormatter" should {
-    val testBand = "RGB"
-    val bandTextFormatter = RasterBandTextFormatter(testBand)
-
-    "map a string band name to the same string representation" in {
-      val bandText = bandTextFormatter.formatString(null, null, null)
-      bandText must be equalTo "RGB"
-    }
-  }
-
 }
